+++ concordium-consensus/src/Concordium/KonsensusV1/TestMonad.hs
@@ -81,6 +81,8 @@ data TestState pv = TestState
     { -- |The 'SkovData'.
       _tsSkovData :: !(SkovData pv),
       -- |The pending timers.
+      -- The 'Integer' key is a handle for the 'Timer' associated
+      -- with the 'TimerMonad'.
       _tsPendingTimers :: !(Map.Map Integer (Timeout, TestMonad pv ()))
     }
 
@@ -102,7 +104,7 @@ data TestEvent (pv :: ProtocolVersion)
       OnPendingLive
     deriving (Eq, Show)
 
--- |Write event monoid. This is simply a list of events.
+-- |List of events generated during a test run.
 type TestWrite pv = [TestEvent pv]
 
 -- |The internals of the test monad.